Saving an intermediate result: - Login/SignUp, user profile, Workspaces work properly with API - New mechanism of maintain and control user session - Minor images preview UI changes - Started to implement UI to find users and share spaces

+++ presentation/src/main/java/com/netcracker_study_autumn_2020/presentation/ui/activity/CreateWorkspaceActivity.java
@@ -13,7 +13,7 @@ public class CreateWorkspaceActivity extends BaseActivity {
 
     private CreateWorkspaceFragment createWorkspaceFragment;
 
-    public static Intent getCallingIntent(Context context, int userId) {
+    public static Intent getCallingIntent(Context context, long userId) {
         Intent intent = new Intent(context, CreateWorkspaceActivity.class);
         intent.putExtra("CURRENT_USER_ID", userId);
         return intent;
